last updates

+++ src/main/java/com/antecsis/dto/DashboardVentasDTO.java
@@ -1,5 +1,7 @@
 package com.antecsis.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -9,5 +11,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class DashboardVentasDTO {
     private Long totalVentas;
-    private Double montoTotal;
+    private BigDecimal montoTotal;
 }
